Added function get_input() to validate user input.
This function accepts a prompt as an argument, which it passes
to raw_input().  It checks the return value of that function to
ensure that the value is either an empty string, or a positive
integer.  In all other cases, it prints a brief error message
and calls itself again.

All uses of raw_input() in the code have been changed to use
get_input() now.

A few minor changes to the code as part of clean-up.

Finally, in def GetAcres(), an optional argument and some VERY
basic logic has been added as an experiment in preparation for
unit testing.